Extracting :home:localization with Stream features text formatting
