+++ app/src/main/java/com/javernaut/whatthecodec/home/ui/subtitle/SubtitlePage.kt
@@ -1,19 +1,15 @@
 package com.javernaut.whatthecodec.home.ui.subtitle
 
-import android.content.res.Resources
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
-import com.javernaut.whatthecodec.R
+import com.javernaut.whatthecodec.feature.home.stream.toDisplayableStreamFeature
 import com.javernaut.whatthecodec.feature.settings.api.content.SubtitleStreamFeature
 import com.javernaut.whatthecodec.home.presentation.model.SubtitlesPage
-import com.javernaut.whatthecodec.home.ui.stream.DisplayableStreamFeature
 import com.javernaut.whatthecodec.home.ui.stream.SimplePage
 import com.javernaut.whatthecodec.home.ui.stream.StreamFeaturesGrid
 import io.github.javernaut.mediafile.SubtitleStream
-import io.github.javernaut.mediafile.displayable.displayableLanguage
-import io.github.javernaut.mediafile.displayable.getDisplayableDisposition
 
 @Composable
 fun SubtitlePage(
@@ -48,25 +44,3 @@ private fun SubtitleCardContent(
         modifier
     )
 }
-
-fun SubtitleStreamFeature.toDisplayableStreamFeature(
-    stream: SubtitleStream,
-    resources: Resources
-) = when (this) {
-    SubtitleStreamFeature.Codec -> stream.basicInfo.codecName
-    SubtitleStreamFeature.Language -> stream.basicInfo.displayableLanguage
-    SubtitleStreamFeature.Disposition -> stream.basicInfo.getDisplayableDisposition(resources)
-}?.let {
-    DisplayableStreamFeature(
-        name = resources.getString(displayableResource),
-        value = it
-    )
-}
-
-
-val SubtitleStreamFeature.displayableResource: Int
-    get() = when (this) {
-        SubtitleStreamFeature.Codec -> R.string.page_subtitle_codec_name
-        SubtitleStreamFeature.Language -> R.string.page_stream_language
-        SubtitleStreamFeature.Disposition -> R.string.page_stream_disposition
-    }
