+++ app/src/main/java/com/javernaut/whatthecodec/home/ui/video/VideoPage.kt
@@ -1,6 +1,5 @@
 package com.javernaut.whatthecodec.home.ui.video
 
-import android.content.res.Resources
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -12,17 +11,14 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.javernaut.whatthecodec.R
+import com.javernaut.whatthecodec.feature.home.stream.toDisplayableStreamFeature
 import com.javernaut.whatthecodec.feature.settings.api.content.VideoStreamFeature
 import com.javernaut.whatthecodec.home.presentation.model.VideoPage
-import com.javernaut.whatthecodec.home.ui.stream.DisplayableStreamFeature
 import com.javernaut.whatthecodec.home.ui.stream.StreamCard
 import com.javernaut.whatthecodec.home.ui.stream.StreamFeatureItem
 import com.javernaut.whatthecodec.home.ui.stream.StreamFeaturesGrid
 import com.javernaut.whatthecodec.home.ui.stream.makeCardTitle
 import io.github.javernaut.mediafile.VideoStream
-import io.github.javernaut.mediafile.displayable.displayableLanguage
-import io.github.javernaut.mediafile.displayable.getDisplayableDisposition
-import io.github.javernaut.mediafile.displayable.toDisplayable
 
 @Composable
 fun VideoPage(
@@ -119,32 +115,3 @@ private fun VideoStream(
         )
     }
 }
-
-fun VideoStreamFeature.toDisplayableStreamFeature(
-    stream: VideoStream,
-    resources: Resources
-) = when (this) {
-    VideoStreamFeature.Codec -> stream.basicInfo.codecName
-    VideoStreamFeature.Bitrate -> stream.bitRate.toDisplayable(resources)
-    VideoStreamFeature.FrameRate -> stream.frameRate.toDisplayable(resources)
-    VideoStreamFeature.FrameWidth -> stream.frameWidth.toString()
-    VideoStreamFeature.FrameHeight -> stream.frameHeight.toString()
-    VideoStreamFeature.Language -> stream.basicInfo.displayableLanguage
-    VideoStreamFeature.Disposition -> stream.basicInfo.getDisplayableDisposition(resources)
-}?.let {
-    DisplayableStreamFeature(
-        name = resources.getString(displayableResource),
-        value = it
-    )
-}
-
-val VideoStreamFeature.displayableResource: Int
-    get() = when (this) {
-        VideoStreamFeature.Codec -> R.string.page_video_codec_name
-        VideoStreamFeature.Bitrate -> R.string.page_video_bit_rate
-        VideoStreamFeature.FrameRate -> R.string.page_video_frame_rate
-        VideoStreamFeature.FrameWidth -> R.string.page_video_frame_width
-        VideoStreamFeature.FrameHeight -> R.string.page_video_frame_height
-        VideoStreamFeature.Language -> R.string.page_stream_language
-        VideoStreamFeature.Disposition -> R.string.page_stream_disposition
-    }
