+++ app/src/main/res/values-uk/strings.xml
@@ -13,24 +13,9 @@
     <string name="info_container">Контейнер</string>
     <string name="info_file_format">Формат файлу</string>
     <string name="info_protocol_title">Використаний протокол</string>
-    <string name="info_bit_rate">Бітрейт</string>
     <string name="page_video_preview_missing_decoder">Нема декодеру\nЗображення недоступно</string>
     <string name="page_video_preview_frame_decoding_error">Не вдалося декодувати кадр</string>
-    <string name="page_video_codec_name">Відео кодек</string>
-    <string name="page_video_bit_rate">@string/info_bit_rate</string>
-    <string name="page_video_frame_rate">Частота кадрів</string>
-    <string name="page_video_frame_width">Ширина</string>
-    <string name="page_video_frame_height">Висота</string>
     <string name="page_stream_title_prefix">Поток #</string>
-    <string name="page_audio_codec_name">Аудіо кодек</string>
-    <string name="page_audio_bit_rate">@string/info_bit_rate</string>
-    <string name="page_audio_channels">Канали</string>
-    <string name="page_audio_channel_layout">Схема каналів</string>
-    <string name="page_audio_sample_rate">Частота дискретизації</string>
-    <string name="page_audio_sample_format">Формат дискретизації</string>
-    <string name="page_subtitle_codec_name">Кодек субтитрів</string>
-    <string name="page_stream_language">Мова</string>
-    <string name="page_stream_disposition">Диспозиція</string>
     <string name="page_stream_no_info">Нема інформації для показу</string>
     <string name="stream_copy">Скопіювати</string>
     <string name="stream_text_copied_pattern" formatted="false">\"%s\" скопійовано в буфер обміну</string>
