+++ app/src/main/res/values/strings.xml
@@ -13,24 +13,9 @@
     <string name="info_container">Container</string>
     <string name="info_file_format">File format</string>
     <string name="info_protocol_title">Used protocol</string>
-    <string name="info_bit_rate">Bit rate</string>
     <string name="page_video_preview_missing_decoder">Missing decoder\nPreview is unavailable</string>
     <string name="page_video_preview_frame_decoding_error">Couldn\'t decode this frame</string>
-    <string name="page_video_codec_name">Video codec</string>
-    <string name="page_video_bit_rate">@string/info_bit_rate</string>
-    <string name="page_video_frame_rate">Frame rate</string>
-    <string name="page_video_frame_width">Width</string>
-    <string name="page_video_frame_height">Height</string>
     <string name="page_stream_title_prefix">Stream #</string>
-    <string name="page_audio_codec_name">Audio codec</string>
-    <string name="page_audio_bit_rate">@string/info_bit_rate</string>
-    <string name="page_audio_channels">Channels</string>
-    <string name="page_audio_channel_layout">Channel layout</string>
-    <string name="page_audio_sample_rate">Sample rate</string>
-    <string name="page_audio_sample_format">Sample format</string>
-    <string name="page_subtitle_codec_name">Subtitle codec</string>
-    <string name="page_stream_language">Language</string>
-    <string name="page_stream_disposition">Disposition</string>
     <string name="page_stream_no_info">No info to display</string>
     <string name="stream_copy">Copy</string>
     <string name="stream_text_copied_pattern" formatted="false">\"%s\" copied to clipboard</string>
