+++ features/home/localization/src/main/kotlin/com/javernaut/whatthecodec/feature/home/stream/DisplayableStreamFeature.kt
@@ -1,4 +1,4 @@
-package com.javernaut.whatthecodec.home.ui.stream
+package com.javernaut.whatthecodec.feature.home.stream
 
 data class DisplayableStreamFeature(
     val name: String,
