+++ features/home/localization/src/main/res/values/strings.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="info_bit_rate">Bit rate</string>
+    <string name="page_video_codec_name">Video codec</string>
+    <string name="page_video_bit_rate">@string/info_bit_rate</string>
+    <string name="page_video_frame_rate">Frame rate</string>
+    <string name="page_video_frame_width">Width</string>
+    <string name="page_video_frame_height">Height</string>
+    <string name="page_audio_codec_name">Audio codec</string>
+    <string name="page_audio_bit_rate">@string/info_bit_rate</string>
+    <string name="page_audio_channels">Channels</string>
+    <string name="page_audio_channel_layout">Channel layout</string>
+    <string name="page_audio_sample_rate">Sample rate</string>
+    <string name="page_audio_sample_format">Sample format</string>
+    <string name="page_subtitle_codec_name">Subtitle codec</string>
+    <string name="page_stream_language">Language</string>
+    <string name="page_stream_disposition">Disposition</string>
+</resources>
