+++ settings.gradle.kts
@@ -30,3 +30,4 @@ rootProject.name = "WhatTheCodec"
 include(":app")
 include(":features:settings:api")
 include(":features:settings:data")
+include(":features:home:localization")
